requirements.txt
appdirs==1.4.0
aubio==0.4.4
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
numpy==1.12.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pydub==0.18.0
pyparsing==2.1.10
six==1.10.0
Werkzeug==0.11.15
